+++ app/src/main/java/com/freshdigitable/udonroad/listitem/QuotedStatusView.java
@@ -14,11 +14,13 @@
  * limitations under the License.
  */
 
-package com.freshdigitable.udonroad;
+package com.freshdigitable.udonroad.listitem;
 
 import android.content.Context;
 import android.util.AttributeSet;
 import android.view.View;
+
+import com.freshdigitable.udonroad.R;
 
 import twitter4j.Status;
 import twitter4j.URLEntity;
