+++ app/src/main/res/layout/fragment_status_detail.xml
@@ -27,7 +27,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             >
-            <com.freshdigitable.udonroad.StatusDetailView
+            <com.freshdigitable.udonroad.listitem.StatusDetailView
                 android:id="@+id/status_view"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
